#1074: Splitted up standard SLC tests
