+++ test_container/tests/test/standard-flavor/7.1/python/available-packages-azure-sdk.py
