+++ test_container/tests/test/standard-flavor/all/python/connect_to_exasol_with_websocket_api.py
